Rename & move AugmentedRecord (now AugmentedModel)
